refactor: move code: bcast_xxx moved to epaxos/replication/ test utils such as foo_inst! moved to epaxos/testutil.rs
