+++ components/epaxos/src/replication/mod.rs
@@ -7,5 +7,11 @@ pub use errors::*;
 mod replication;
 pub use replication::*;
 
+mod broadcast;
+pub use broadcast::*;
+
 #[cfg(test)]
 mod test_hdlreply;
+
+#[cfg(test)]
+mod test_broadcast;
